refactored into unified ParseEnvelope api
