+++ src/inscriptions.rs
@@ -5,11 +5,10 @@ use tag::Tag;
 pub(crate) use self::media::Media;
 
 pub use self::{
-  embedding::ParsedEmbedding, envelope::Envelope, envelope::ParsedEnvelope, envelope::RawEnvelope,
-  inscription::Inscription, inscription_id::InscriptionId,
+  envelope::Envelope, envelope::ParsedEnvelope, envelope::RawEnvelope, inscription::Inscription,
+  inscription_id::InscriptionId,
 };
 
-mod embedding;
 mod envelope;
 mod inscription;
 pub(crate) mod inscription_id;
