+++ src/lib.rs
@@ -97,9 +97,7 @@ pub use self::{
   chain::Chain,
   fee_rate::FeeRate,
   index::{Index, RuneEntry},
-  inscriptions::{
-    Envelope, Inscription, InscriptionId, ParsedEmbedding, ParsedEnvelope, RawEnvelope,
-  },
+  inscriptions::{Envelope, Inscription, InscriptionId, ParsedEnvelope, RawEnvelope},
   object::Object,
   options::Options,
   wallet::transaction_builder::{Target, TransactionBuilder},
