+++ tests/decode.rs
@@ -66,6 +66,7 @@ fn from_file() {
         offset: 0,
         pushnum: false,
         stutter: false,
+        annex: false,
       }],
       runestone: Some(Artifact::Runestone(Runestone::default())),
     },
@@ -89,6 +90,7 @@ fn from_stdin() {
         offset: 0,
         pushnum: false,
         stutter: false,
+        annex: false,
       }],
       runestone: Some(Artifact::Runestone(Runestone::default())),
     },
@@ -121,6 +123,7 @@ fn from_core() {
         offset: 0,
         pushnum: false,
         stutter: false,
+        annex: false,
       }],
       runestone: None,
     },
